requirements.txt
fastapi
uvicorn[standard]
torch
ftfy
regex
tqdm
Pillow
python-multipart
git+https://github.com/openai/CLIP.git
